refactor(settings): extract add models dialog
