+++ packages/web-design/src/components/fleet-pi/pi/config-panel/sections/model-defaults-section.tsx
@@ -1,19 +1,11 @@
-import { Plus, RefreshCw, Search, Trash2 } from "lucide-react"
+import { Plus, Search, Trash2 } from "lucide-react"
 import { useMemo, useState } from "react"
 import { Button } from "../../../../button"
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "../../../../dialog"
 import {
   InputGroup,
   InputGroupAddon,
   InputGroupInput,
 } from "../../../../input-group"
-import { Spinner } from "../../../../spinner"
 import { cn } from "../../../../../lib/utils"
 import { ItemRow } from "../../../primitives/item-row"
 import {
@@ -21,6 +13,7 @@ import {
   SettingsPane,
 } from "../../../primitives/settings-pane"
 import { HIT_AREA_EXPAND_DENSE_CLASS } from "../../../styles/tokens"
+import { AddModelsDialog } from "./add-models-dialog"
 import { isModelEnabled } from "../shared/model-patterns"
 import {
   ProviderBrandIcon,
@@ -61,11 +54,8 @@ export function ModelDefaultsSection({
   settingsLoading: boolean
 }) {
   const [addOpen, setAddOpen] = useState(false)
-  const [addFilter, setAddFilter] = useState("")
-  const [selectedKeys, setSelectedKeys] = useState<Set<string>>(new Set())
   const disabled = !draft || settingsLoading
   const normalizedFilter = modelFilter.trim().toLowerCase()
-  const normalizedAddFilter = addFilter.trim().toLowerCase()
 
   const configuredProviderIds = useMemo(() => {
     // Trust the server (prime-agent-driven /api/chat/providers). Anything it
@@ -107,39 +97,6 @@ export function ModelDefaultsSection({
     return [...groups.entries()]
   }, [listedModels])
 
-  const candidateModels = useMemo(() => {
-    return modelOptions.filter((model) => {
-      if (!configuredProviderIds.has(model.provider)) return false
-      if (isModelEnabled(model, draft?.enabledModels)) return false
-      if (!normalizedAddFilter) return true
-      return [model.name, model.modelId, model.provider, model.id]
-        .join(" ")
-        .toLowerCase()
-        .includes(normalizedAddFilter)
-    })
-  }, [
-    configuredProviderIds,
-    draft?.enabledModels,
-    modelOptions,
-    normalizedAddFilter,
-  ])
-
-  const groupedCandidates = useMemo(() => {
-    const groups = new Map<string, Array<ConfigModelInfo>>()
-    for (const model of candidateModels) {
-      const existing = groups.get(model.provider)
-      if (existing) existing.push(model)
-      else groups.set(model.provider, [model])
-    }
-    return [...groups.entries()]
-  }, [candidateModels])
-
-  const discoverableProviders = useMemo(() => {
-    return [...configuredProviderIds].sort((a, b) =>
-      formatProviderLabel(a).localeCompare(formatProviderLabel(b))
-    )
-  }, [configuredProviderIds])
-
   const emptyMessage = (() => {
     if (modelOptions.length === 0) {
       return "No models discovered yet. Configure a provider, then add models."
@@ -154,27 +111,7 @@ export function ModelDefaultsSection({
   })()
 
   const openAddDialog = () => {
-    setSelectedKeys(new Set())
-    setAddFilter("")
     setAddOpen(true)
-  }
-
-  const toggleSelected = (model: ConfigModelInfo) => {
-    setSelectedKeys((current) => {
-      const next = new Set(current)
-      if (next.has(model.id)) next.delete(model.id)
-      else next.add(model.id)
-      return next
-    })
-  }
-
-  const confirmAdd = () => {
-    const selected = candidateModels.filter((model) =>
-      selectedKeys.has(model.id)
-    )
-    if (selected.length > 0) onAddModels(selected)
-    setAddOpen(false)
-    setSelectedKeys(new Set())
   }
 
   return (
@@ -286,135 +223,17 @@ export function ModelDefaultsSection({
         )}
       </div>
 
-      <Dialog open={addOpen} onOpenChange={setAddOpen}>
-        <DialogContent className="max-w-lg">
-          <DialogHeader>
-            <DialogTitle>Add models</DialogTitle>
-            <DialogDescription>
-              Pick models discovered from your configured providers. For OpenAI
-              Chat Completions, refresh runs{" "}
-              <span className="font-medium">GET {"{baseUrl}/models"}</span>.
-            </DialogDescription>
-          </DialogHeader>
-
-          <div className="flex flex-wrap gap-1.5">
-            {discoverableProviders.map((providerId) => {
-              const discovering = discoveringProviderId === providerId
-              return (
-                <Button
-                  key={providerId}
-                  type="button"
-                  variant="outline"
-                  size="sm"
-                  disabled={discoveringProviderId !== null}
-                  onClick={() => {
-                    void onDiscoverProvider(providerId)
-                  }}
-                >
-                  {discovering ? (
-                    <Spinner data-icon="inline-start" />
-                  ) : (
-                    <RefreshCw data-icon="inline-start" />
-                  )}
-                  {formatProviderLabel(providerId)}
-                </Button>
-              )
-            })}
-          </div>
-
-          <InputGroup>
-            <InputGroupAddon align="inline-start">
-              <Search />
-            </InputGroupAddon>
-            <InputGroupInput
-              aria-label="Search discovered models"
-              value={addFilter}
-              onChange={(event) => setAddFilter(event.target.value)}
-              placeholder="Search discovered models…"
-            />
-          </InputGroup>
-
-          <div className="max-h-72 overflow-y-auto">
-            {groupedCandidates.length === 0 ? (
-              <p className="py-6 text-center text-xs text-pretty text-muted-foreground">
-                {candidateModels.length === 0 && !normalizedAddFilter
-                  ? "All discovered models are already in your list, or none are available yet."
-                  : "No models match your search."}
-              </p>
-            ) : (
-              <div className="flex flex-col gap-3">
-                {groupedCandidates.map(([provider, models]) => (
-                  <div key={provider} className="flex flex-col gap-1">
-                    <div className="px-2 text-[11px] font-medium tracking-wide text-muted-foreground uppercase">
-                      {formatProviderLabel(provider)}
-                    </div>
-                    {models.map((model) => {
-                      const checked = selectedKeys.has(model.id)
-                      return (
-                        <button
-                          key={model.id}
-                          type="button"
-                          className={cn(
-                            "flex w-full items-center gap-3 rounded-lg px-2 py-2 text-left transition-[background-color,transform] duration-150",
-                            "hover:bg-muted/60 focus-visible:ring-2 focus-visible:ring-ring focus-visible:outline-none active:scale-[0.96]",
-                            checked && "bg-muted/70"
-                          )}
-                          aria-pressed={checked}
-                          onClick={() => toggleSelected(model)}
-                        >
-                          <div className="flex size-8 shrink-0 items-center justify-center rounded-[4px] border border-border/40 bg-background/60">
-                            <ProviderBrandIcon
-                              provider={model.provider}
-                              className="text-foreground/70"
-                            />
-                          </div>
-                          <div className="min-w-0 flex-1">
-                            <div className="truncate text-sm font-medium">
-                              {model.name}
-                            </div>
-                            <div className="truncate text-xs text-muted-foreground">
-                              {model.modelId}
-                            </div>
-                          </div>
-                          <span
-                            className={cn(
-                              "size-4 rounded border border-border/60",
-                              checked && "border-foreground bg-foreground"
-                            )}
-                            aria-hidden
-                          />
-                        </button>
-                      )
-                    })}
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-
-          <div className="flex items-center justify-end gap-1.5">
-            <Button
-              type="button"
-              variant="ghost"
-              size="sm"
-              onClick={() => setAddOpen(false)}
-            >
-              Cancel
-            </Button>
-            <Button
-              type="button"
-              size="sm"
-              disabled={selectedKeys.size === 0}
-              onClick={confirmAdd}
-            >
-              Add{" "}
-              {selectedKeys.size > 0 ? (
-                <span className="tabular-nums">({selectedKeys.size})</span>
-              ) : null}
-            </Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      {addOpen ? (
+        <AddModelsDialog
+          configuredProviderIds={configuredProviderIds}
+          discoveringProviderId={discoveringProviderId}
+          enabledModels={draft?.enabledModels}
+          modelOptions={modelOptions}
+          onAddModels={onAddModels}
+          onDiscoverProvider={onDiscoverProvider}
+          onOpenChange={setAddOpen}
+        />
+      ) : null}
     </SettingsPane>
   )
 }
